refactor: UserNavigation, ManagerNavigation을 components로 이동한다
