+++ front-end/src/components/ManagerNavigation/ManagerNavigation.stories.tsx
