+++ front-end/src/components/ManagerNavigation/ManagerNavigation.styled.ts
