+++ front-end/src/components/index.ts
@@ -10,6 +10,7 @@ export { default as Input } from "./Input/Input";
 export { default as Layout } from "./Layout/Layout";
 export { default as Loading } from "./Loading/Loading";
 export { default as Logo } from "./Logo/Logo";
+export { default as ManagerNavigation } from "./ManagerNavigation/ManagerNavigation";
 export { default as Portal } from "./Portal/Portal";
 export { default as Radio } from "./Radio/Radio";
 export { default as RadioGroup } from "./RadioGroup/RadioGroup";
@@ -19,4 +20,5 @@ export { default as Table } from "./Table/Table";
 export { default as Toast } from "./Toast/Toast";
 export { default as ToastProvider, useToast } from "./Toast/ToastProvider";
 export { default as Text } from "./Text/Text";
+export { default as UserNavigation } from "./UserNavigation/UserNavigation";
 export { default as VerticalBox } from "./VerticalBox/VerticalBox";
