--- front-end/src/domains/Manager.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as ManagerNavigation } from "./ManagerNavigation/ManagerNavigation";
