--- front-end/src/domains/User.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as UserNavigation } from "./UserNavigation/UserNavigation";
