+++ front-end/src/pages/GpuServerViewDetail/GpuServerViewDetail.tsx
@@ -1,7 +1,5 @@
 import { useMyInfo, useServerId } from "../../hooks";
-import { Layout, Header, SubHeader } from "../../components";
-import { ManagerNavigation } from "../../domains/Manager";
-import { UserNavigation } from "../../domains/User";
+import { Layout, Header, SubHeader, ManagerNavigation, UserNavigation } from "../../components";
 import { GpuServerDetail } from "../../features/gpuServer";
 
 const GpuServerViewDetail = () => {
